==> scripts/02-download_data.R <==
#### Preamble ####
# Purpose: Downloads and saves the data from opendatatoronto
# Author: Angel Xu
# Date: 22 November 2024
# Contact: anjojoo.xu@mail.utoronto.ca
# License: MIT
# Pre-requisites: None
# Any other information needed? None


#### Workspace setup ####
library(opendatatoronto)
library(tidyverse)
library(dplyr)
library(here)
library(lintr)
library(styler)

#### Download data ####
# get package
package <- show_package("260e1356-dce6-48e2-afa0-e71d70cd6406")
package

# get all resources for this package
resources <- list_package_resources("260e1356-dce6-48e2-afa0-e71d70cd6406")

# identify datastore resources; by default, Toronto Open Data sets datastore resource format to CSV for non-geospatial and GeoJSON for geospatial resources
datastore_resources <- filter(resources, tolower(format) %in% c('csv', 'geojson'))

# load the first datastore resource as a sample
data <- filter(datastore_resources, row_number()==1) %>% get_resource()
data



#### Save data ####
# change the_raw_data to whatever name you assigned when you downloaded it.
write_csv(data, "data/01-raw_data/active_applications.csv") 

         
#### Lint and style the code ####
lint(filename = here("scripts/02-download_data.R"))
style_file(path = here("scripts/02-download_data.R"))
